Strictly js hubot
Transpiled hubot all down to js. Going to start messing around in here
and figure out what is actually goin on in hubot to get a better
understanding.

So far there are three EventEmitters in hubot: Robut, Adapter and Brain.
Need to figure out a way to incorporate some sort of state dialog
management.